// src/types/api.ts
export interface components {
    schemas: {
        /** MemberLoginRequest */
        MemberLoginRequest: {
            /** @description 비밀번호 기반 인증용 해시값 */
            passwordAuthHash: string;
            /** @description 로그인 시도할 회원 id (이메일 등) */
            loginId: string;
        };
        /** VoidResponse */
        VoidResponse: {
            /** @description 응답 상태 코드 (ex: ok) */
            code: string;
            /** @description 결과 메시지 (ex: ok) */
            data: string;
            /** @description 응답 메시지 */
            message: string;
            /** @description HTTP 상태 코드 */
            status: number;
        };
        /** MemberCreateRequest */
        MemberCreateRequest: {
            /** @description KDF 파라미터 */
            kdfParam: {
                /** @description 출력 키 길이 bytes 단위 (최소 32) */
                keyLength: number;
                /** @description 메모리 사용량 KB 단위 (최소 19456) */
                memory: number;
                /** @description 병렬 처리 수 (최소 1) */
                parallelism: number;
                /** @description 반복 횟수 (최소 2) */
                iterations: number;
                /** @description KDF 알고리즘 (예: ARGON2ID) */
                algorithm: string;
            };
            /** @description 비밀번호 분실 대비 복구키로 암호화된 DEK */
            recoveryEncryptedDek: string;
            /** @description 비밀번호 기반 인증용 해시값 */
            passwordAuthHash: string;
            /** @description 회원 로그인 id (이메일 등) */
            loginId: string;
            /** @description 사용자의 메인 KEK로 암호화된 데이터 암호화 키 (DEK) */
            passwordEncryptedDek: string;
            /** @description 복구키 해시 생성에 사용된 고유 솔트값 */
            recoverySalt: string;
            /** @description 복구키 기반 인증용 해시값 */
            recoveryAuthHash: string;
            /** @description 비밀번호 해시 생성에 사용된 고유 솔트값 */
            passwordSalt: string;
        };
        /** LoginInitResponse */
        LoginInitResponse: {
            /** @description 코드 */
            code: string;
            /** @description 응답 데이터 */
            data: {
                /** @description 클라이언트 KDF 파라미터 정보 */
                kdfParam: {
                    /** @description 출력 키 길이 bytes 단위 (최소 32) */
                    keyLength: number;
                    /** @description 메모리 사용량 KB 단위 (최소 19456) */
                    memory: number;
                    /** @description 병렬 처리 수 (최소 1) */
                    parallelism: number;
                    /** @description 반복 횟수 (최소 2) */
                    iterations: number;
                    /** @description KDF 알고리즘 (예: argon2id) */
                    algorithm: string;
                };
                /** @description 해시 생성에 사용된 고유 솔트값 */
                passwordSalt: string;
            };
            /** @description 메시지 */
            message: string;
            /** @description 상태 */
            status: number;
        };
        /** NodeTreeResponse */
        NodeTreeResponse: {
            /** @description 응답 상태 코드 (ex: ok) */
            code: string;
            /** @description 노드 트리 리스트 */
            data: {
                /** @description 부모 노드의 식별자 ID (최상위 루트 폴더인 경우 null) */
                nodeParentId?: number | null;
                /** @description 종단간 암호화(E2EE)가 적용된 노드 제목 */
                encryptedTitle: string;
                /** @description 노드의 종류 (FOLDER 또는 FILE) */
                nodeType: string;
                /** @description 노드의 고유 식별자(ID) */
                nodeId: number;
            }[];
            /** @description 응답 메시지 */
            message: string;
            /** @description HTTP 상태 코드 */
            status: number;
        };
        /** MemberLoginInitRequest */
        MemberLoginInitRequest: {
            /** @description 로그인 시도할 회원 id (이메일 등) */
            loginId: string;
        };
        /** MemberRecoveryRequest */
        MemberRecoveryRequest: {
            /** @description 새로운 비밀번호 기반 인증 해시 */
            newPasswordAuthHash: string;
            /** @description 회원 로그인 id */
            loginId: string;
            /** @description 새로운 비밀번호용 솔트 */
            newPasswordSalt: string;
            /** @description 새로 적용할 KDF 파라미터 */
            newKdfParam: {
                /** @description 키 길이 */
                keyLength: number;
                /** @description 메모리 사용량 */
                memory: number;
                /** @description 병렬 처리 수 */
                parallelism: number;
                /** @description 반복 횟수 */
                iterations: number;
                /** @description KDF 알고리즘 */
                algorithm: string;
            };
            /** @description 새로운 비밀번호(KEK)로 암호화된 DEK */
            newPasswordEncryptedDek: string;
            /** @description 기존 복구키 기반 인증 해시 */
            recoveryAuthHash: string;
        };
        /** MemberMeResponse */
        MemberMeResponse: {
            /** @description 코드 */
            code: string;
            /** @description 응답 데이터 */
            data: {
                /** @description 회원 로그인 id */
                loginId: string;
                /** @description 가입 권한 - ACCOUNT_AUTH, PASSKEY */
                authMethod?: string | null;
                /** @description 회원 ID */
                memberId: number;
            };
            /** @description 에러 메시지 */
            message: string;
            /** @description 상태 */
            status: number;
        };
        /** NodeCreateRequest */
        NodeCreateRequest: {
            /** @description 위치 지정을 위한 다음(하단) 노드 ID (없을 시 null) */
            nextNodeId?: number | null;
            /** @description 위치 지정을 위한 이전(상단) 노드 ID (없을 시 null) */
            prevNodeId?: number | null;
            /** @description 종단간 암호화(E2EE)가 적용된 노드 제목 */
            encryptedTitle: string;
            /** @description 생성할 노드의 부모 ID (최상위 루트에 생성 시 null) */
            parentNodeId?: number | null;
            /** @description 생성할 노드의 종류 (FOLDER 또는 FILE) */
            nodeType: string;
        };
        /** MemberRecoveryInitResponse */
        MemberRecoveryInitResponse: {
            /** @description 코드 */
            code: string;
            /** @description 응답 데이터 */
            data: {
                /** @description 클라이언트 KDF 파라미터 정보 */
                kdfParam: {
                    /** @description 출력 키 길이 bytes 단위 (최소 32) */
                    keyLength: number;
                    /** @description 메모리 사용량 KB 단위 (최소 19456) */
                    memory: number;
                    /** @description 병렬 처리 수 (최소 1) */
                    parallelism: number;
                    /** @description 반복 횟수 (최소 2) */
                    iterations: number;
                    /** @description KDF 알고리즘 (예: argon2id) */
                    algorithm: string;
                };
                /** @description 복구 키에 따른 암호화된 DEK */
                recoveryEncryptedDek: string;
                /** @description 복구 키 해시 생성에 사용된 고유 솔트값 */
                recoverySalt: string;
            };
            /** @description 메시지 */
            message: string;
            /** @description 상태 */
            status: number;
        };
        /** NodeContentCreateRequest */
        NodeContentCreateRequest: {
            /** @description 종단간 암호화(E2EE)가 적용된 본문 내용 */
            encryptedContents: (Record<string, never> | boolean | string | number)[];
            /** @description 위치 지정을 위한 다음(하단) 블록 ID (없을 시 null) */
            nextNodeContentId?: number | null;
            /** @description 위치 지정을 위한 이전(상단) 블록 ID (없을 시 null) */
            prevNodeContentId?: number | null;
        };
        /** NodeFileResponse */
        NodeFileResponse: {
            /** @description 응답 상태 코드 (ex: ok) */
            code: string;
            /** @description 노드(파일)의 상세 내용 블록 리스트 */
            data: {
                /** @description 상세 내용 블록의 고유 식별자(ID) */
                nodeContentId: number;
                /** @description 종단간 암호화(E2EE)가 적용된 본문 내용 */
                encryptedContent: string;
            }[];
            /** @description 응답 메시지 */
            message: string;
            /** @description HTTP 상태 코드 */
            status: number;
        };
        "api-v1-exception-1466924442": {
            /** @description 코드 */
            code: string;
            /** @description 응답 데이터 */
            data?: string | null;
            /** @description API 접근 권한이 없을때 생기는 오류 */
            message: string;
            /** @description 상태 */
            status: number;
        };
        /** NodeContentCreateResponse */
        NodeContentCreateResponse: {
            /** @description 응답 상태 코드 (ex: ok) */
            code: string;
            /** @description 생성된 노드 컨텐츠 응답 데이터 리스트 */
            data: {
                /** @description 새로 생성된 노드 컨텐츠의 고유 ID */
                nodeContentId: number;
            }[];
            /** @description 응답 메시지 */
            message: string;
            /** @description HTTP 상태 코드 */
            status: number;
        };
        /** MemberLoginResponse */
        MemberLoginResponse: {
            /** @description 코드 */
            code: string;
            /** @description 응답 데이터 */
            data: {
                /** @description 암호화된 마스터 키 */
                passwordEncryptedDek: string;
                /** @description JWT 액세스 토큰 */
                accessToken: string;
            };
            /** @description 에러 메시지 */
            message: string;
            /** @description 상태 */
            status: number;
        };
        /** NodeTitleUpdateRequest */
        NodeTitleUpdateRequest: {
            /** @description 수정할 종단간 암호화(E2EE)가 적용된 노드 제목 */
            encryptedTitle: string;
        };
        /** NodeContentBatchUpdateRequest */
        NodeContentBatchUpdateRequest: {
            /** @description 수정할 노드 내용(블록) 요청 목록 */
            requests: {
                /** @description 수정할 대상 노드 내용(블록)의 고유 ID */
                nodeContentId: number;
                /** @description 수정할 종단간 암호화(E2EE)가 적용된 본문 내용 */
                encryptedContent: string;
            }[];
        };
        /** NodeMoveRequest */
        NodeMoveRequest: {
            /** @description 위치 지정을 위한 다음(하단) 노드 ID (없을 시 null) */
            nextNodeId?: number | null;
            /** @description 위치 지정을 위한 이전(상단) 노드 ID (없을 시 null) */
            prevNodeId?: number | null;
            /** @description 이동할 위치의 부모 노드 ID (최상위 루트로 이동 시 null) */
            parentNodeId?: number | null;
        };
        /** MemberRecoveryInitRequest */
        MemberRecoveryInitRequest: {
            /** @description 계정 복구를 시도할 회원 id (이메일 등) */
            loginId: string;
        };
        /** MemberRecoveryResponse */
        MemberRecoveryResponse: {
            /** @description 코드 */
            code: string;
            /** @description 응답 데이터 */
            data: {
                /** @description 새로 발급된 JWT 액세스 토큰 */
                accessToken: string;
            };
            /** @description 메시지 */
            message: string;
            /** @description 상태 */
            status: number;
        };
        /** NodeContentDeleteRequest */
        NodeContentDeleteRequest: {
            /** @description 수정할 노드 내용(블록) 요청 목록 id */
            nodeContentIds: (Record<string, never> | boolean | string | number)[];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}